test(encryption tab): move test about cached secrets in `EncryptionUserSettingsTab-test.tsx`
